+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - [Storage] Added MinIO storage backend
 - [Core] Allow to pass function args as part of the invocation payload in FaaS backends
+- [Core] Optimize call_async() calls with an internal function cache system
 
 ### Changed
 - [AWS Lambda] Use layer from Klayers API for pre-compiled Amazon Linux numpy binaries
@@ -17,6 +18,10 @@
 - [Core] Fixed customized_runtime feature
 - [Core] Fixed get_result() execution after wait() when throw_except is set to False
 - [Core] Fixed internal executions
+- [Core] Fixed 'lithops storage list' CLI when a bucket is empty
+
+### Deleted
+- [Core] Deleted invoke_pool_threads param from map and map_reduce calls. Now it must be set at backend level in config
 
 
 ## [v2.3.5]